refactor: xmodule/raw_module.py -> xmodule/raw_block.py
