+++ xmodule/word_cloud_module.py
@@ -16,7 +16,7 @@ from web_fragments.fragment import Fragment
 from xblock.core import XBlock
 from xblock.fields import Boolean, Dict, Integer, List, Scope, String
 from xmodule.editing_block import EditingMixin
-from xmodule.raw_module import EmptyDataRawMixin
+from xmodule.raw_block import EmptyDataRawMixin
 from xmodule.util.xmodule_django import add_webpack_to_fragment
 from xmodule.xml_module import XmlMixin
 from xmodule.x_module import (
